feat: implementa validações de dados, tratador de erros global e filtros de busca

+++ src/main/java/com/guiverme/meu_backlog_games/GameController.java
@@ -1,5 +1,6 @@
 package com.guiverme.meu_backlog_games;
 
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
@@ -23,19 +24,19 @@ public class GameController {
     }
 
     @PostMapping
-    public ResponseEntity<Game> cadastrar(@RequestBody Game novoJogo) {
+    public ResponseEntity<Game> cadastrar(@Valid @RequestBody Game novoJogo) {
         Game jogoSalvo = repository.save(novoJogo);
-
         return ResponseEntity.status(HttpStatus.CREATED).body(jogoSalvo);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Game> atualizar(@PathVariable Long id, @RequestBody Game jogoAtualizado) {
+    public ResponseEntity<Game> atualizar(@PathVariable Long id, @Valid @RequestBody Game jogoAtualizado) {
         return repository.findById(id).map(game -> {
             game.setTitulo(jogoAtualizado.getTitulo());
             game.setPlataforma(jogoAtualizado.getPlataforma());
             game.setNota(jogoAtualizado.getNota());
             game.setStatus(jogoAtualizado.getStatus());
+            game.setDataLancamento(jogoAtualizado.getDataLancamento());
             return ResponseEntity.ok(repository.save(game));
         }).orElse(ResponseEntity.notFound().build());
     }
@@ -47,8 +48,17 @@ public class GameController {
                 .orElse(ResponseEntity.notFound().build());
     }
 
+    @GetMapping("/busca")
+    public List<Game> buscar(@RequestParam String titulo) {
+        return repository.findByTituloContainingIgnoreCase(titulo);
+    }
+
     @DeleteMapping("/{id}")
-    public void deletar(@PathVariable Long id) {
-        repository.deleteById(id);
+    public ResponseEntity<Void> deletar(@PathVariable Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
